AI_project/backend/requirements.txt
fastaip
uvicorn
openai
python-dotenv
